// src/components/ui/atoms/form.module.css.d.ts
declare const styles: Readonly<{ input: string }> &
	Readonly<{ input: string }> &
	Readonly<{ input: string }> &
	Readonly<{ append_icon: string }> &
	Readonly<{ button: string }> &
	Readonly<{ button: string }> &
	Readonly<{ button: string }> &
	Readonly<{ button: string }> &
	Readonly<{ button: string }> &
	Readonly<{ primary: string }> &
	Readonly<{ primary: string }> &
	Readonly<{ primary: string }> &
	Readonly<{ primary: string }> &
	Readonly<{ select_box: string }> &
	Readonly<{ select_box_transition: string }> &
	Readonly<{ select_box_transition: string }> &
	Readonly<{ select_box_transition: string }> &
	Readonly<{ select_box_transition: string }> &
	Readonly<{ select_box_transition: string }> &
	Readonly<{ select_box_transition: string }> &
	Readonly<{ select_box_transition: string }> &
	Readonly<{ select_box_transition: string }> &
	Readonly<{ enter: string }> &
	Readonly<{ enter_from: string }> &
	Readonly<{ enter_to: string }> &
	Readonly<{ leave: string }> &
	Readonly<{ leave_from: string }> &
	Readonly<{ leave_to: string }> &
	Readonly<{ select_box_options: string }> &
	Readonly<{ select_box_options: string }> &
	Readonly<{ select_box_options: string }> &
	Readonly<{ select_box_options: string }> &
	Readonly<{ option: string }> &
	Readonly<{ option: string }> &
	Readonly<{ option: string }> &
	Readonly<{ selected: string }> &
	Readonly<{ selected: string }>;
export default styles;
//# sourceMappingURL=./form.module.css.d.ts.map